feat: pending actions table on dashboard (#1993)
<img width="2192" alt="image"
src="https://github.com/user-attachments/assets/af9d178e-b531-42a4-bee5-c78ca263d62b"
/>

<img width="2192" alt="image"
src="https://github.com/user-attachments/assets/58c4e954-5709-4886-9b76-0210768ea431"
/>


## Summary by Sourcery

Migrate actions table functionality to a server+client split by turning
ActionsTable and ActionsDropdownTable into async server components that
fetch data, extract client-side rendering logic into ActionsClientTable
and ActionsDropdownClientTable, wrap them in Suspense with skeleton
loaders, and add the pending actions view to the admin dashboard

New Features:
- Convert ActionsTable and ActionsDropdownTable into async server
components that fetch action data internally
- Introduce ActionsClientTable and ActionsDropdownClientTable to handle
rendering and empty states on the client
- Add a pending actions table section to the admin assets dashboard page

Enhancements:
- Remove manual data filtering and UI state from server components by
delegating to client tables
- Wrap action table components with React Suspense and DataTableSkeleton
for loading states
- Unify data fetching via getActionsList and getUser across pages and
components
- Simplify page components by removing inline getActionsList calls
