+++ kit/dapp/src/components/blocks/actions-table/actions-dropdown-table.tsx
@@ -1,103 +1,35 @@
-"use client";
-
-import { Badge } from "@/components/ui/badge";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import type {
-  Action,
-  ActionStatus,
-} from "@/lib/queries/actions/actions-schema";
-import { useTranslations } from "next-intl";
-import { useState } from "react";
+import { getUser } from "@/lib/auth/utils";
+import { getActionsList } from "@/lib/queries/actions/actions-list";
+import { ActionType } from "@/lib/queries/actions/actions-schema";
+import type { Address } from "viem";
 import type { DataTablePaginationOptions } from "../data-table/data-table-pagination";
 import type { DataTableToolbarOptions } from "../data-table/data-table-toolbar";
-import { ActionsTable } from "./actions-table";
+import { ActionsDropdownClientTable } from "./actions-dropdown-client-table";
 
 interface ActionsDropdownTableProps {
-  actions: Action[];
-
+  type: ActionType;
+  targetAddress?: Address;
   toolbar?: DataTableToolbarOptions;
   pagination?: DataTablePaginationOptions;
 }
 
-export function ActionsDropdownTable({
-  actions,
+export async function ActionsDropdownTable({
+  type,
+  targetAddress,
   toolbar,
   pagination,
 }: ActionsDropdownTableProps) {
-  const t = useTranslations("actions");
-  const [status, setStatus] = useState<ActionStatus>("PENDING");
-  const counts = actions.reduce(
-    (acc, action) => {
-      acc[action.status] = (acc[action.status] ?? 0) + 1;
-      return acc;
-    },
-    {} as Record<ActionStatus, number>
-  );
-  const statusOptions = [
-    {
-      value: "PENDING",
-      label: t("status.PENDING"),
-      count: counts.PENDING,
-    },
-    {
-      value: "UPCOMING",
-      label: t("status.UPCOMING"),
-      count: counts.UPCOMING,
-    },
-    {
-      value: "COMPLETED",
-      label: t("status.COMPLETED"),
-      count: counts.COMPLETED,
-    },
-    {
-      value: "EXPIRED",
-      label: t("status.EXPIRED"),
-      count: counts.EXPIRED,
-    },
-  ];
-  const handleStatusChange = (newStatus: ActionStatus) => {
-    setStatus(newStatus);
-  };
-
+  const user = await getUser();
+  const actionsList = await getActionsList({
+    userAddress: user.wallet,
+    type,
+    targetAddress,
+  });
   return (
-    <div className="space-y-4">
-      <div className="flex justify-start">
-        <Select
-          value={status}
-          onValueChange={(value) => handleStatusChange(value as ActionStatus)}
-        >
-          <SelectTrigger className="w-[200px]">
-            <SelectValue defaultValue={t("status.PENDING")} />
-          </SelectTrigger>
-          <SelectContent>
-            {statusOptions.map((option) => (
-              <SelectItem key={option.value} value={option.value}>
-                <div className="flex items-center justify-between w-full">
-                  <span>{option.label}</span>
-                  {option.count > 0 && (
-                    <Badge variant="outline" className="ml-2 border-card">
-                      {option.count}
-                    </Badge>
-                  )}
-                </div>
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-      </div>
-
-      <ActionsTable
-        status={status}
-        actions={actions}
-        toolbar={toolbar}
-        pagination={pagination}
-      />
-    </div>
+    <ActionsDropdownClientTable
+      actions={actionsList}
+      toolbar={toolbar}
+      pagination={pagination}
+    />
   );
 }
